+++ spec/melody/package_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 # Check installed packages
-%w(chef git).each do |pkg|
+%w(chef git foodcritic trceroute).each do |pkg|
   describe package(pkg) do
     it { should be_installed }
   end